# docs/API_SPEC.yaml
openapi: 3.1.0
...
info:
  title: Dr. Tahir LMS API
  version: 2.0.0-redesign
  description: >
    Backend for tahirelshazli.com. This version reconciles the existing API with
    the Claude Design product redesign. See docs/API_GAP_ANALYSIS.md for the
    route-by-route disposition and docs/AUTHORIZATION_MODEL.md for the rules
    every `x-roles` annotation below is shorthand for.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.tahirelshazli.com
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation failed. `message` is an array of constraint strings.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: >
        Authenticated but the wrong ROLE for this action. Never used for
        out-of-scope resources — those are 404.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: >
        The resource does not exist, OR exists outside the caller's scope. The
        body is identical in both cases, deliberately.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The action contradicts current state (e.g. already sent).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        error: { type: string }
    Id:
      type: string
      maxLength: 64
      pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$'
...
    Course:
      type: object
      required:
        [id, slug, isPublished, title, description, teacherName,
         sequentialLockEnabled, modules]
      properties:
        id:          { $ref: '#/components/schemas/Id' }
        slug:        { type: string, pattern: '^[a-z0-9-]+$', maxLength: 120 }
        isPublished: { type: boolean }
        title:       { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        thumbnailUrl: { type: [string, 'null'], maxLength: 2048 }
        teacherName: { type: string, maxLength: 160 }
        sequentialLockEnabled: { type: boolean }
        modules:     { type: array, items: { type: object } }
    CourseWrite:
      type: object
      required: [title, description, slug, teacherName]
      properties:
        title:       { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        slug:        { type: string, pattern: '^[a-z0-9-]+$', maxLength: 120 }
        teacherName: { type: string, minLength: 1, maxLength: 160 }
        thumbnailUrl: { type: [string, 'null'], maxLength: 2048 }
        sequentialLockEnabled: { type: boolean }
        # Absent means DRAFT. A course that published itself on creation would
        # put an empty outline on the public site.
        isPublished: { type: boolean }

    # Every field optional. `thumbnailUrl` is the only one where `null` is a
    # value ("clear it") rather than a 400 - the rest are NOT NULL columns.
    CoursePatch:
      type: object
      properties:
        title:       { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        slug:        { type: string, pattern: '^[a-z0-9-]+$', maxLength: 120 }
        teacherName: { type: string, minLength: 1, maxLength: 160 }
        thumbnailUrl: { type: [string, 'null'], maxLength: 2048 }
        sequentialLockEnabled: { type: boolean }
        isPublished: { type: boolean }
...
paths:
...
  /staff/announcements/reach:
    get:
      tags: [announcements]
      summary: How many students an audience currently reaches
      description: >
        Resolved live. The audience is never stored as a frozen recipient list —
        a stored list silently misses anyone who joins after drafting.
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: audience, in: query, required: true,
            schema: { type: string, pattern: '^(all_students|all_tas|course:[A-Za-z0-9_-]+|group:[A-Za-z0-9_-]+)$' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { reach: { type: integer } }
...
  /staff/courses/{courseId}/announcements:
    get:
      tags: [announcements]
      summary: List course announcements
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '200': { description: OK }
    post:
      tags: [announcements]
      summary: Create course draft
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '201': { description: Created }
  /staff/courses/{courseId}/announcements/{id}:
    patch:
      tags: [announcements]
      summary: Update course draft
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '200': { description: Updated }
    delete:
      tags: [announcements]
      summary: Delete course draft
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '204': { description: Deleted }
  /staff/groups/{groupId}/announcements:
    get:
      tags: [announcements]
      summary: List group announcements
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: groupId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '200': { description: OK }
    post:
      tags: [announcements]
      summary: Create group draft
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: groupId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '201': { description: Created }
  /staff/groups/{groupId}/announcements/{id}:
    patch:
      tags: [announcements]
      summary: Update group draft
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: groupId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '200': { description: Updated }
    delete:
      tags: [announcements]
      summary: Delete group draft
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: groupId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '204': { description: Deleted }
...
  /admin/courses:
    post:
      tags: [settings]
      summary: Create a course
      description: >
        Created as a DRAFT unless `isPublished` says otherwise, with an empty
        outline - modules and lessons are authored afterwards. Unscoped:
        creating a course is not a question an assistant is asked.
      x-roles: [teacher, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseWrite' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
      x-audit: course.created
  /admin/courses/{courseId}:
    get:
      tags: [settings]
      summary: Course detail for editing
      description: >
        Course detail for staff editing. Unscoped: teacher and admin only.
      x-roles: [teacher, admin]
      parameters:
        - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [settings]
      summary: Edit a course
      description: >
        Partial. The outline is not writable here. 409 on a slug another course
        already holds.
      x-roles: [teacher, admin]
      parameters:
        - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CoursePatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
      x-audit: course.updated
...